requirements.txt
Flask==2.2.3
Flask-PyMongo==2.3.0
PyJWT==2.6.0
requests==2.28.2
flask-sqlalchemy==3.0.3
flask-migrate==4.0.4
python-decouple==3.8
pudb==2022.1.3